Add route tracking db registration. Add map route details on finished events. Fix map polylines

+++ app/src/main/java/com/muei/apm/runtrack/data/models/Location.kt
@@ -3,7 +3,7 @@ package com.muei.apm.runtrack.data.models
 import java.util.Date
 
 data class Location(
-        var latitude: Long,
-        var longitude: Long,
+        var latitude: Double,
+        var longitude: Double,
         var date: Date = Date()
 )
